requirements.txt
about-time==4.2.1
alive-progress==3.1.5
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==23.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
executing==2.0.1
grapheme==0.6.0
h11==0.14.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
JPype1==1.5.0
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mypy==1.6.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.31.0
selenium==4.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
tabula-py==2.9.3
tornado==6.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.18
urllib3-secure-extra==0.1.0
wcwidth==0.2.13
webdriver-manager==4.0.1
wsproto==1.2.0
